refactored code and created validation.py
